Made some changes and new additions.
Took Daniels upgrades, made the code structure cleaner and more organised and also started on a tutorial level to introduce to placers. Daniel and I thought it would be funny to have the texts speak in a annoyed tone seeing as people have played hundreds of platformers before. Somethings are not going to be commited seeing as it would create conflict with what is found in Daniels-Upgrades, they will be commited once this branch and Daniels-Upgrades branch are merged either Friday at club or sunday night.